feat: Async FIPA REQUEST PROTOCOL
The idea is to apply an asynchronous-like notation using generators in
order to construct routines and procedures in a sequential fashion,
abstracting the internal logic of FIPA interaction protocols.

When an async transaction resumes after being "yielded", it still has access to local
variables
defined beforehand.

This will facilitate multiple concurrent transactions for a single agent,
instead of relying on class attributes to keep track of states and
context.

An example is also available in examples folder.