feat: Automatic exception handler in generator
It won't be necessary to catch every exception. For an asynchronous
request that only expects one response, the generator becomes more
similar to python coroutines.

When multiple responses are expected (for example, an AGREE followed by
an INFORM), an infinity yield loop should still be placed, catching the
FipaProtocolComplete exception at the end to break the loop.